frontend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic-settings
PyMuPDF
faiss-cpu
httpx
pydantic
python-docx
